add: main 컴포넌트 분리, home-user contents 작성가 이미지, 닉네임 표시
